requirements.txt

vaderSentiment==3.3.2
emoji==2.2.0
deep-translator
psycopg2-binary
qrcode[pil]
whitenoise>=6.0.0
asgiref==3.8.1
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
chardet==3.0.4
charset-normalizer==3.3.2
cryptography==43.0.0
defusedxml==0.8.0rc2
dj-rest-auth==6.0.0
Django==5.1
django-allauth==64.0.0
django-cors-headers==4.4.0
django-oauth-toolkit==3.0.1
django-rest-knox==5.0.1
django-rest-passwordreset==1.4.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-api-core==2.23.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-language==2.15.1
googleapis-common-protos==1.66.0
googlemaps==4.10.0
googletrans==3.0.0
grpcio==1.68.1
grpcio-status==1.68.1
gunicorn==22.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.12.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
jwcrypto==1.5.6
langdetect==1.0.9
oauth2client==4.1.3
oauthlib==3.2.2
packaging==24.1
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.2
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
psycopg2-binary
cairosvg
python-decouple
pyotp
pytz
django-extensions
aiohttp
python-dateutil
google-generativeai

